qagarden_agents/pyproject.toml
[project]
name = "qagarden_agents"
version = "0.1.0"
description = "AgentStack SDK wrappers for QA Garden agents"
authors = [
    { name = "QA Garden Team" }
]
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "agentstack-sdk",
    "httpx",
    "a2a-sdk",
]